refactor: moved linksData to linksData.ts file
