+++ src/components/molecules/Footer/index.tsx
@@ -6,43 +6,7 @@ import styles from './Footer.module.css';
 
 import { currentYear } from '@/lib/dates/dates';
 
-const socialMedias = [
-  {
-    id: '1',
-    src: '/images/social-icons/instagram.svg',
-    url: 'https://www.google.com',
-  },
-  {
-    id: '2',
-    src: '/images/social-icons/twitter.svg',
-    url: 'https://www.google.com',
-  },
-  {
-    id: '3',
-    src: '/images/social-icons/facebook.svg',
-    url: 'https://facebook.com/m3ntorship',
-  },
-  {
-    id: '4',
-    src: '/images/social-icons/linkedin.svg',
-    url: 'https://linkedin.com/company/m3ntorship/',
-  },
-];
-
-const links = {
-  about: [
-    { id: '1', name: 'Story', route: '/' },
-    { id: '2', name: 'Programs', route: '/' },
-    { id: '3', name: 'Alumni', route: '/' },
-    { id: '4', name: 'Mentors', route: '/' },
-    { id: '5', name: 'Sponsors', route: '/' },
-    { id: '6', name: 'Contact', route: '/' },
-  ],
-  conditions: [
-    { id: '1', name: 'Terms of use', route: '/' },
-    { id: '2', name: 'Privacy policy', route: '/' },
-  ],
-};
+import { links, socialMedias } from '@/components/molecules/Footer/linksData';
 
 export default function Footer(): ReactElement {
   return (
